Separated into two libraries
