+++ src/CMakeLists.txt
@@ -1,0 +1,4 @@
+cmake_minimum_required(VERSION 3.12)
+
+add_subdirectory(core_internal)
+add_subdirectory(core)
