+++ src/core/communication/communicator_manager.cpp
