+++ src/core/communication/dummy/dummy_communicator.cpp
