+++ src/core/compute/checks.cpp
