+++ src/core/compute/cpu/cpu_device.cpp
