+++ src/core/compute/cpu/cpu_device_backend.cpp
@@ -3,6 +3,7 @@
 #include <xmipp4/core/compute/cpu/cpu_device_backend.hpp>
 
 #include <xmipp4/core/compute/cpu/cpu_device.hpp>
+#include <xmipp4/core/compute/device_manager.hpp>
 #include <xmipp4/core/core_version.hpp>
 #include <xmipp4/core/system/host.hpp>
