+++ src/core/compute/device_manager.cpp
