+++ src/core/core_version.cpp
