+++ src/core/memory/aligned_alloc.cpp
