+++ src/core/plugin_manager.cpp
