+++ src/core/system/dynamic_library.cpp
@@ -2,7 +2,7 @@
 
 #include <xmipp4/core/system/dynamic_library.hpp>
 
-#include "dynamic_library_handle.hpp"
+#include <system/dynamic_library_handle.hpp>
 
 #include <sstream>
