+++ src/core_internal/plugin_loader.cpp
