+++ src/core_internal/system/dynamic_library_handle_windows.inl
