+++ src/core_internal/system/memory_mapped_file_handle.hpp
